backend/requirements.txt
asyncio
flask
fastapi
uvicorn
sqlalchemy
asyncpg
passlib
PyJWT
kafka-python
